project files updated from angular 1 to 13

+++ src/app/06-services/todos.component.ts
@@ -1,25 +1,30 @@
+import { Component } from '@angular/core';
+import { TodoService } from './todo.service';
 
-import { TodoService } from './todo.service'
-
-export class TodosComponent { 
+@Component({
+  selector: 'app-todos',
+  templateUrl: './todo.component.html',
+  styleUrls: [],
+})
+export class TodosComponent {
   todos: any[] = [];
-  message; 
+  message: any;
 
   constructor(private service: TodoService) {}
 
-  ngOnInit() { 
-    this.service.getTodos().subscribe(t => this.todos = t);
+  ngOnInit() {
+    this.service.getTodos().subscribe((t) => (this.todos = t));
   }
 
-  add() { 
+  add() {
     var newTodo = { title: '... ' };
-    this.service.add(newTodo).subscribe(
-      t => this.todos.push(t),
-      err => this.message = err);
+    this.service.add(newTodo).subscribe({
+      next: (t) => this.todos.push(t),
+      error: (err) => (this.message = err),
+    });
   }
 
-  delete(id) {
-    if (confirm('Are you sure?'))
-      this.service.delete(id).subscribe();
-  }  
+  delete(id: any) {
+    if (confirm('Are you sure?')) this.service.delete(id).subscribe();
+  }
 }
